requirements.txt
pydantic>=2.0
openai>=1.0
hypothesis>=6.0
python-dotenv>=1.0
